refactor: scaffolding
