+++ src/api/pageview.ts
@@ -1,4 +1,4 @@
-import { type Route, getSettings } from "../settings";
+import { type Route, getSettings } from "../core/settings";
 import type { GtagConfigParams } from "../types/gtag";
 import query from "./query";
