+++ src/api/query.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "../settings";
+import { getSettings } from "../core/settings";
 import type { Gtag } from "../types/gtag";
 import { isServer } from "../utils";
