+++ src/core/add-configuration.ts
@@ -1,8 +1,8 @@
-import { consentDeniedAll, consentGrantedAll } from "./api/consent";
-import linker from "./api/linker";
-import query from "./api/query";
+import { consentDeniedAll, consentGrantedAll } from "../api/consent";
+import linker from "../api/linker";
+import query from "../api/query";
+import type { GtagConfig } from "../types/gtag";
 import { getSettings } from "./settings";
-import type { GtagConfig } from "./types/gtag";
 
 function mergeDefaults(config: GtagConfig = {}): GtagConfig {
   return {
