+++ src/core/settings.test.ts
