+++ src/core/track-route.test.ts
@@ -1,9 +1,9 @@
 import { type Router, createRouter, createWebHistory } from "vue-router";
-import query from "./api/query";
+import query from "../api/query";
 import { resetSettings, updateSettings } from "./settings";
 import trackRoute from "./track-route";
 
-vi.mock("./api/query");
+vi.mock("../api/query");
 
 describe("track-route", () => {
   let router: Router;
