+++ src/core/track-route.ts
@@ -1,5 +1,5 @@
-import pageview from "./api/pageview";
-import screenview from "./api/screenview";
+import pageview from "../api/pageview";
+import screenview from "../api/screenview";
 import { type PageTrackerParams, type Route, getSettings } from "./settings";
 
 function isRouteExcluded(route: Route): boolean {
